Move basic step definitions out of 'DataCoverageStepDefinitions'
And put them into 'IndividualEnrollmentStepDefinitions', a new
class for basic navigation and data entry step definitions for
individual providers, where they can be used for more than just
data coverage checks.

Issue #518 Implement accessibility checking in CI
