+++ psm-app/integration-tests/src/test/java/gov/medicaid/features/enrollment/steps/DataCoverageStepDefinitions.java
@@ -1,6 +1,5 @@
 package gov.medicaid.features.enrollment.steps;
 
-import cucumber.api.java.en.Given;
 import cucumber.api.java.en.Then;
 import cucumber.api.java.en.When;
 import gov.medicaid.features.enrollment.ui.OrganizationInfoPage;
@@ -18,55 +17,6 @@ public class DataCoverageStepDefinitions {
     GeneralSteps generalSteps;
 
     private OrganizationInfoPage organizationInfoPage;
-
-    @Given("I am on the individual provider license info page")
-    public void enter_individual_provider_license_info_page() {
-        navigateToIndividualProviderLicensePage();
-    }
-
-    @Given("^I am on the organizational provider license info page$")
-    public void enter_organizational_provider_license_info_page() {
-        navigateToOrganizationalProviderLicensePage();
-    }
-
-    private void navigateToIndividualProviderLicensePage() {
-        generalSteps.loginAsProvider();
-        enrollmentSteps.createEnrollment();
-        enrollmentSteps.selectIndividualProviderType();
-        enrollmentSteps.enterIndividualPersonalInfo();
-        enrollmentSteps.advanceFromIndividualPersonalInfoToLicenseInfo();
-        enrollmentSteps.enterNotAProviderAtPublicHealthServiceIndianHospital();
-    }
-
-    private void navigateToOrganizationalProviderLicensePage() {
-        generalSteps.loginAsProvider();
-        enrollmentSteps.createEnrollment();
-        enrollmentSteps.selectOrganizationalProviderType();
-        enrollmentSteps.enterOrganizationInfo();
-        enrollmentSteps.enterContactInfo();
-        enrollmentSteps.advanceFromOrganizationInfoToLicenseInfo();
-    }
-
-    @Given("^I am on the individual provider practice info page$")
-    public void enter_individual_provider_practice_info_page() throws IOException {
-        navigateToProviderPracticePage();
-    }
-
-    private void navigateToProviderPracticePage() throws IOException {
-        navigateToIndividualProviderLicensePage();
-        enrollmentSteps.enterNotAProviderAtPublicHealthServiceIndianHospital();
-        enrollmentSteps.enterLicenseInfo();
-        enrollmentSteps.uploadLicense();
-        enrollmentSteps.advanceFromIndividualLicenseInfoToPracticeInfo();
-    }
-
-    @Given("^I am on the individual provider statement page$")
-    public void enter_individual_provider_statement_page() throws IOException {
-        navigateToProviderPracticePage();
-        enrollmentSteps.enterIndividualPrivatePracticeInfo();
-        enrollmentSteps.advanceFromIndividualPracticeInfoToSummaryPage();
-        enrollmentSteps.advanceFromIndividualSummaryToProviderStatementPage();
-    }
 
     @When("^I enter valid license information$")
     public void enter_valid_license_information() {
@@ -109,11 +59,6 @@ public class DataCoverageStepDefinitions {
     @Then("^I should be asked to enter Medicaid number$")
     public void i_should_be_asked_to_enter_Medicaid() {
         organizationInfoPage.verifyMedicaidNumberAccepted();
-    }
-
-    @When("^I move to the personal info page$")
-    public void i_move_to_the_personal_info_page() {
-        enrollmentSteps.selectIndividualProviderType();
     }
 
     @Then("^I should be asked to enter Applicant Name, Contact Person$")
